requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.0.1
certifi==2024.8.30
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
ecdsa==0.19.0
fastapi==0.114.2
filelock==3.16.1
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
identify==2.6.1
idna==3.10
iniconfig==2.0.0
Mako==1.3.5
MarkupSafe==2.1.5
nodeenv==1.9.1
packaging==24.1
passlib==1.7.4
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==3.8.0
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
virtualenv==20.26.4
